pyproject.toml
# TBD
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-feature-gen"
version = "0.1.8"
description = "Feature discovery and generation utilities"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Julija Jeršova" }]
license = { text = "MIT" }
dependencies = [
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2",
  "numpy>=1.26",
  "Pillow>=10",
  "opencv-python>=4.8",
  "ffmpeg-python>=0.2",
  "pandas>=2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["src/tests"]
addopts = "--cov=llm_feature_gen --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["llm_feature_gen"]

[tool.coverage.report]
fail_under = 100
show_missing = true
skip_empty = true

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"llm_feature_gen" = ["prompts/*.txt"]

# Belt-and-suspenders: don't bundle any non-code files from these folders
[tool.setuptools.exclude-package-data]
"*" = [
  "discover_images/*",
  "discover_tabular/*",
  "discover_texts/*",
  "discover_videos/*",
  "images/*",
  "outputs/*",
  "tabular/*",
  "texts/*",
  "videos/*",
]
